fix(Player e Ball) : correção de padrões
